requirements.txt
apscheduler==3.10.4
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
tweepy==4.15.0
google-generativeai==0.3.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
flask==2.3.3
